sheet revamp v1.5
I forgot to test filters and sorting on last revamp. This final update finally puts sheet UI to the same standard as icon UI. Refreshing the list works the same as icons. Interactions such as filter, sort, and level editing are butter smooth now as the sheet is practically destroyed to refresh just like icons